requirements.txt
Flask~=3.1.2
Flask-CORS~=6.0.1
requests~=2.32.5
beautifulsoup4~=4.14.2
PyYAML~=6.0.3
bcrypt~=5.0.0